% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataVision.R
\name{dataVision}
\alias{dataVision}
\title{dataVision}
\usage{
dataVision(x, y, df, verbose = TRUE, na.rm = FALSE)
}
\arguments{
\item{na.rm}{}
}
\value{
a plot visualizing your x and y data from your specified dataset
}
\description{
dataVision
}
\examples{
plotFunction('Sepal.Width', 'Petal.Width', iris)

}
